Updated imports due to interface and struct moves
